refactor(app): extract app.py into focused modules
Decompose 955-line God file into focused modules:
- translation.py: LLM translation, language detection, sentence splitting
- services.py: OpenAI/AWS session management, health check server
- websocket_handler.py: WebSocket server with decomposed handler functions
- components/scroll_lock.html: extracted inline CSS/JS

Replace global WEBSOCKET_PORT with dict ref for thread-safe sharing.
app.py reduced to ~213 lines as slim UI controller.

Closes #9
